Rewrite code editor UI and make sample apps handle their own static asset requests

+++ browserver.js
@@ -29,42 +29,81 @@ kernel.use({http: 0, html: 0, database: 0, xhr: 0, string: 0}, function(o, proxy
           });
         }
         return o.database.get('modules', function(modules) {
+          Object.keys(modules).forEach(function(name, i) {
+            modules[name] = {code: modules[name], running: !!apps[name]};
+          });
           response.end(o.html.markup([
             {'!doctype': {html: null}},
             {html: [
               {head: [
                 {title: 'Browserver'},
-                {link: {rel: 'shortcut icon', href: 'http://localhost:8001/icon.png'}}
+                {meta: {charset: 'utf-8'}},
+                {link: {rel: 'shortcut icon', href: '/icon.png'}},
+                {link: {rel: 'stylesheet', href: '/codemirror/codemirror.css'}},
+                {link: {rel: 'stylesheet', href: '/browserver.css'}}
               ]},
               {body: [
-                {input: {type: 'text', placeholder: 'Module name', id: 'name'}},
-                {button: {children: 'Add', disabled: 'disabled', id: 'add'}},
-                {form: {action: '/', method: 'post', children: [
-                  {ul: {id: 'modules', children: Object.keys(modules).map(function(module) {
-                    return {li: [
-                      {a: {href: '/'+encodeURIComponent(module), target: '_blank', children: module}}, ' ',
-                      {button: {type: 'submit', name: 'module', value: module, children: apps[module] ? 'Stop' : 'Run'}}
-                    ]};
-                  })}}
-                ]}},
-                {script: {src: 'http://localhost:8001/kernel.js'}},
-                {script: {src: 'http://localhost:8001/modules/html.js'}},
+                {script: {src: '/kernel.js'}},
+                {script: {src: '/modules/html.js'}},
+                {script: {src: '/modules/xhr.js'}},
+                {script: {src: '/codemirror/codemirror.js'}},
+                {script: {src: '/codemirror/javascript.js'}},
+                {script: {src: '/codemirror/matchbrackets.js'}},
+                {script: {src: '/codemirror/match-highlighter.js'}},
                 {script: function(m) {
-                  if (!m) return Object.keys(modules);
-                  kernel.use({html: 0}, function(o) {
-                    var add = document.getElementById('add'),
-                        name = document.getElementById('name'),
-                        list = document.getElementById('modules');
-                    name.onkeyup = function() {
-                      add.disabled = !this.value;
+                  if (!m) return modules;
+                  kernel.use({html: 0, xhr: 0}, function(o) {
+                    var name, add, list, tab, active = Object.keys(m)[0], entry = function(name) {
+                      var encoded = encodeURIComponent(name),
+                          module = m[name];
+                      return {li: function(e) {
+                        if (name == active)
+                          (tab = e).className = 'active';
+                        e.onclick = function() {
+                          if (tab) tab.className = '';
+                          (tab = this).className = 'active';
+                          editor.setValue(module.code);
+                        };
+                        return [
+                          name,
+                          {button: {children: module.running ? 'Stop' : 'Run', onclick: function(e) {
+                            e.stopPropagation();
+                            this.disabled = true;
+                            o.xhr('/', {method: 'POST', headers: {'Content-Type': 'application/x-www-form-urlencoded'}, data: 'module='+encoded}, function() {
+                              e.target.disabled = false;
+                              e.target.textContent = (module.running = !module.running) ? 'Stop' : 'Run';
+                            });
+                          }}}
+                        ];
+                      }};
                     };
-                    add.onclick = function() {
-                      if (~m.indexOf(name.value)) return alert('Module name already exists');
-                      m.push(name.value);
-                      o.html.dom({li: [
-                        {a: {href: '/'+encodeURIComponent(name.value), target: '_blank', children: name.value}}, ' ',
-                        {button: {type: 'submit', name: 'module', value: name.value, children: 'Run'}}
-                      ]}, list);
+                    o.html.dom([
+                      {div: {id: 'nav', children: [
+                        {div: {id: 'module', children: [
+                          {input: {type: 'text', placeholder: 'New Module', onkeyup: function() { add.disabled = !this.value; }, children: function(e) { name = e; }}},
+                          {button: {disabled: 'disabled', children: function(e) { add = e; return 'Add'; }, onclick: function() {
+                            if (m[name.value]) return alert('Module name already exists');
+                            m[name.value] = {code: ''};
+                            tab.className = '';
+                            active = name.value;
+                            o.html.dom(entry(name.value), list);
+                            name.value = '';
+                          }}}
+                        ]}},
+                        {ul: function(e) { list = e; return Object.keys(m).map(entry); }}
+                      ]}},
+                      {div: {id: 'editor', children: [
+                        {textarea: function(e) { code = e; return active ? m[active].code : ''; }}
+                      ]}}
+                    ], document.body);
+                    var editor = CodeMirror.fromTextArea(code, {
+                      lineNumbers: true,
+                      matchBrackets: true,
+                      highlightSelectionMatches: true
+                    });
+                    CodeMirror.commands.save = function() {
+                      if (!active) return;
+                      o.xhr('/'+encodeURIComponent(active), {method: 'POST', data: module[active] = editor.getValue()});
                     };
                   });
                 }}
@@ -74,61 +113,26 @@ kernel.use({http: 0, html: 0, database: 0, xhr: 0, string: 0}, function(o, proxy
         });
       }
       
+      if (request.method == 'GET')
+        return o.xhr(location.origin+request.path, {responseType: 'arraybuffer'}, function(e) {
+          if (e.target.status != 200)
+            return response.end('404 Resource not found', null, 404);
+          response.end(e.target.response, {'Content-Type': o.http.getMimeType((request.path.match(/\.([^.]*)$/) || [])[1])});
+        });
+      
       var module = decodeURIComponent(request.path.substr(1)),
           path = 'modules/'+encodeURIComponent(module);
       
       if (request.method == 'DELETE')
         return o.database.delete(path, function() {
-          response.end('Success');
+          response.end('Deleted');
         });
       
       // TODO: validate encoding
       if (request.method == 'POST')
         return o.database.put(path, o.string.fromUTF8Buffer(request.body), function() {
-          response.end('Success', {Location: request.path}, 303);
+          response.end('Saved');
         });
-      
-      o.database.get(path, function(code) {
-        response.end(o.html.markup([
-          {'!doctype': {html: null}},
-          {html: [
-            {head: [
-              {title: 'Browserver'},
-              {meta: {charset: 'utf-8'}},
-              {link: {rel: 'shortcut icon', href: 'http://localhost:8001/icon.png'}},
-              {link: {rel: 'stylesheet', href: 'http://localhost:8001/codemirror/codemirror.css'}},
-              {style: 'body { margin: 0; } .CodeMirror { height: auto; } .CodeMirror-scroll { overflow-x: auto; overflow-y: hidden; }'},
-            ]},
-            {body: [
-              {textarea: code || ''},
-              {script: {src: 'http://localhost:8001/codemirror/codemirror.js'}},
-              {script: {src: 'http://localhost:8001/codemirror/javascript.js'}},
-              {script: {src: 'http://localhost:8001/codemirror/matchbrackets.js'}},
-              {script: {src: 'http://localhost:8001/codemirror/match-highlighter.js'}},
-              {script: function() {
-                CodeMirror.fromTextArea(document.getElementsByTagName('textarea')[0], {
-                  lineNumbers: true,
-                  matchBrackets: true,
-                  highlightSelectionMatches: true
-                });
-                CodeMirror.commands.save = function(cm) {
-                  var request = new XMLHttpRequest();
-                  request.open('POST', location.pathname);
-                  request.send(cm.getValue());
-                };
-              }}
-            ]}
-          ]}
-        ]), {'Content-Type': 'text/html'});
-      });
-    });
-  });
-  
-  o.http.serve({port: 8001}, function(request, response) {
-    o.xhr(location.origin+request.path, {responseType: 'arraybuffer'}, function(e) {
-      if (e.target.status != 200)
-        return response.end('404 Resource not found', null, 404);
-      response.end(e.target.response, {'Content-Type': o.http.getMimeType((request.path.match(/\.([^.]*)$/) || [])[1])});
     });
   });
 });
